I found a few bugs and switched the boundary condition in the brentq optimization to be continuity in Hx a the grounding line.
The code seems to converge using the default paramters (except Nx=513 instead of Nx=1025, I'm testing higher res right now).

However, the noise floor from the solver is quite large, suggesting that the matrix is not well conditioned and this is limiting the accuracy of the final results.  Gunter, maybe you can help me to improve this?